requirements.txt
sentence-transformers
scikit-learn
matplotlib
umap-learn
numpy
seaborn